vscode-vlang: update vls compilation command, fix restart command (#303)
* vscode-vlang: reuse output channel for vls

* vscode-vlang: fix vls not restarting on a user-provided path

* vscode-vlang: fix deactivating and terminating vls

* vscode-vlang: remove subscription on language client

* vscode-vlang: improve vls restart command

* vscode-vlang: transfer vls restart command to its own function

* vscode-vlang: explicitly terminate vls process on "vls.enable" config change

+++ src/langserver.ts
@@ -4,11 +4,11 @@ import fs from 'fs';
 import cp from 'child_process';
 import util from 'util';
 import * as net from 'net';
-import { window, ExtensionContext, workspace, ProgressLocation } from 'vscode';
+import { window, workspace, ProgressLocation } from 'vscode';
 import { LanguageClient, LanguageClientOptions, StreamInfo, ServerOptions } from 'vscode-languageclient/node';
 
 import { getVExecCommand, getWorkspaceConfig } from './utils';
-import { outputChannel } from './status';
+import { outputChannel, vlsOutputChannel } from './status';
 
 const execAsync = util.promisify(cp.exec);
 const mkdirAsync = util.promisify(fs.mkdir);
@@ -18,7 +18,7 @@ const vlsDir = path.join(os.homedir(), '.vls');
 const vlsBin = path.join(vlsDir, 'bin');
 const vexe = getVExecCommand();
 const isWin = process.platform === 'win32';
-export const vlsPath = path.join(vlsBin, isWin ? 'vls.exe' : 'vls');
+export let vlsPath = path.join(vlsBin, isWin ? 'vls.exe' : 'vls');
 export let client: LanguageClient;
 let vlsProcess: cp.ChildProcess;
 let shouldSpawnProcess = true;
@@ -86,7 +86,7 @@ function connectVlsViaTcp(port: number): Promise<StreamInfo> {
 	return Promise.resolve(result);
 }
 
-export function connectVls(pathToVls: string, context: ExtensionContext): void {
+export function connectVls(pathToVls: string): void {
 	const connMode = getWorkspaceConfig().get<string>('vls.connectionMode');
 	const tcpPort = getWorkspaceConfig().get<number>('vls.tcpMode.port');
 
@@ -125,6 +125,10 @@ export function connectVls(pathToVls: string, context: ExtensionContext): void {
 	}
 
 	if (shouldSpawnProcess) {
+		// Kill first the existing VLS process
+		// before launching a new one.
+		terminateVlsProcess();
+
 		console.log('Spawning VLS process...');
 		vlsProcess = cp.spawn(pathToVls.trim(), vlsArgs);
 	}
@@ -139,6 +143,7 @@ export function connectVls(pathToVls: string, context: ExtensionContext): void {
 		synchronize: {
 			fileEvents: workspace.createFileSystemWatcher('**/*.v')
 		},
+		outputChannel: vlsOutputChannel
 	};
 
 	client = new LanguageClient(
@@ -156,28 +161,37 @@ export function connectVls(pathToVls: string, context: ExtensionContext): void {
 			window.setStatusBarMessage('The V language server failed to initialize.', 3000);
 		});
 
-	context.subscriptions.push(client.start());
+	// NOTE: the language client was remove in the context subscriptions
+	// because of it's error-handling behavior which causes the progress/message
+	// box to hang and produce unnecessary errors in the output/devtools log.
+	client.start();
 }
 
-export async function activateVls(context: ExtensionContext): Promise<void> {
+export async function activateVls(): Promise<void> {
 	if (!isVlsEnabled()) return;
 	const customVlsPath = getWorkspaceConfig().get<string>('vls.customPath');
 	if (!customVlsPath) {
 		// if no vls path is given, try to used the installed one or install it.
 		const installed = await checkIsVlsInstalled();
 		if (installed) {
-			connectVls(vlsPath, context);
+			connectVls(vlsPath);
 		}
 	} else {
-		connectVls(customVlsPath, context);
+		// It is very important to set this or start/stopping
+		// the VLS process won't work.
+		vlsPath = customVlsPath;
+		connectVls(customVlsPath);
 	}
 }
 
 export async function deactivateVls(): Promise<void> {
 	if (client && isVlsEnabled()) {
-		await client.stop();
-		if (shouldSpawnProcess && !vlsProcess.killed) {
-			vlsProcess.kill();
-		}
+		return client.stop();
+	}
+}
+
+export function terminateVlsProcess(): void {
+	if (shouldSpawnProcess && typeof vlsProcess != 'undefined' && vlsProcess && !vlsProcess.killed) {
+		vlsProcess.kill();
 	}
 }
